+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/send/payments/CreateInvoiceLink.kt
@@ -1,0 +1,31 @@
+package dev.inmo.tgbotapi.extensions.api.send.payments
+
+import dev.inmo.tgbotapi.bot.TelegramBot
+import dev.inmo.tgbotapi.requests.send.payments.CreateInvoiceLink
+import dev.inmo.tgbotapi.requests.send.payments.SendInvoice
+import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
+import dev.inmo.tgbotapi.types.chat.CommonUser
+import dev.inmo.tgbotapi.types.payments.LabeledPrice
+import dev.inmo.tgbotapi.types.payments.abstracts.Currency
+
+suspend fun TelegramBot.createInvoiceLink(
+    title: String,
+    description: String,
+    payload: String,
+    providerToken: String,
+    currency: Currency,
+    prices: List<LabeledPrice>,
+    maxTipAmount: Int? = null,
+    suggestedTipAmounts: List<Int>? = null,
+    providerData: String? = null,
+    requireName: Boolean = false,
+    requirePhoneNumber: Boolean = false,
+    requireEmail: Boolean = false,
+    requireShippingAddress: Boolean = false,
+    shouldSendPhoneNumberToProvider: Boolean = false,
+    shouldSendEmailToProvider: Boolean = false,
+    priceDependOnShipAddress: Boolean = false
+) = execute(
+    CreateInvoiceLink(title, description, payload, providerToken, currency, prices, maxTipAmount, suggestedTipAmounts ?.sorted(), providerData, requireName, requirePhoneNumber, requireEmail, requireShippingAddress, shouldSendPhoneNumberToProvider, shouldSendEmailToProvider, priceDependOnShipAddress)
+)